tests: Manage exceptions in rdmacm processes
[ Upstream commit a344e834f9affa53d5af7068be555b36e2af8823 ]

Change the rdmacm tests main thread to check the processes' queue
each second for any exceptions.
This will prevent the main thread from waiting for the whole timeout to
be reached before finishing the test in case of a failure on one of the
threads.

In addition, check if there was a SkipTest exception in order to skip
the test instead of failing it.

Fixes: 1be9b6647c49 ("tests: Fix exception handling in rdmacm test")

+++ tests/test_rdmacm.py
@@ -51,8 +51,16 @@ class CMTestCase(RDMATestCase):
                                       'passive':False, **resource_kwargs})
         passive.start()
         active.start()
-        passive.join(15)
-        active.join(15)
+        proc_raised_ex = False
+        for i in range(15):
+            if proc_raised_ex:
+                break
+            for proc in [passive, active]:
+                proc.join(1)
+                if not proc.is_alive() and not self.notifier.empty():
+                    proc_raised_ex = True
+                    break
+
         # If the processes is still alive kill them and fail the test.
         proc_killed = False
         for proc in [passive, active]:
@@ -60,10 +68,18 @@ class CMTestCase(RDMATestCase):
                 proc.terminate()
                 proc_killed = True
         # Check if the test processes raise exceptions.
-        if not self.notifier.empty():
-            res = self.notifier.get()
-            if res is not None:
-                raise PyverbsError(res)
+        proc_res = {}
+        while not self.notifier.empty():
+            res, side = self.notifier.get()
+            proc_res[side] = res
+        for ex in proc_res.values():
+            if isinstance(ex, unittest.case.SkipTest):
+                raise(ex)
+        if proc_res:
+            print(f'Received the following exceptions: {proc_res}')
+            if isinstance(res, Exception):
+                raise(res)
+            raise PyverbsError(res)
         # Raise exeption if the test proceses was terminate.
         if proc_killed:
             raise Exception('RDMA CM test procces is stuck, kill the test')
